use a mixin
